# taskboard/public/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TaskBoard API",
    "version": "1.0.0",
    "description": "A comprehensive API for managing Kanban tasks with integrated MCP (Model Context Protocol) chat functionality.\n\n## Authentication\n\nProtected endpoints require user identification via one of these methods:\n1. **Bearer Token**: `Authorization: Bearer <userId>`\n2. **Query Parameter**: `?userId=<userId>`  \n3. **Request Body**: Include `userId` field in JSON body\n\nThis flexible authentication supports both web applications and MCP integrations.\n\n## Features\n\n- User registration and authentication\n- CRUD operations for Kanban items (todo, doing, done)  \n- MCP-powered AI chat assistant integration\n- Session-based conversation management\n- Tool usage tracking and display"
  },
  "servers": [
    {
      "url": "${API_BASE_URL:-http://localhost:3000}",
      "description": "API server (configurable via API_BASE_URL environment variable)"
    }
  ],
...
  "paths": {
    "/api/chat": {
      "get": {
        "summary": "Chat service health check",
        "description": "Check the health status of the chat service and MCP integration",
        "operationId": "chat_health_check",
        "tags": [
          "Chat"
        ],
        "x-gram": {
          "name": "check_chat_service_health",
          "description": "<context>\nThis endpoint checks if the MCP-powered chat service is healthy and available.\nUse this to verify the chat functionality is working before attempting to send messages.\n</context>\n<prerequisites>\n- userId parameter is required for authentication\n</prerequisites>\n",
          "responseFilterType": "jq"
        },
        "security": [
          {
            "UserIdParam": []
          }
        ],
        "responses": {
          "200": {
            "description": "Chat service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "healthy"
                    },
                    "chat_service": {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "example": "healthy"
                        },
                        "service": {
                          "type": "string",
                          "example": "mcp-chat-microservice"
                        }
                      }
                    }
                  }
                },
                "example": {
                  "status": "healthy",
                  "chat_service": {
                    "status": "healthy",
                    "service": "mcp-chat-microservice"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "message": "Unauthorized"
                }
              }
            }
          },
          "503": {
            "description": "Chat service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "status": "unhealthy",
                  "error": "Chat service unavailable"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Send chat message",
        "description": "Send a message to the MCP-powered AI assistant and receive a response with optional tool usage",
        "operationId": "send_chat_message",
        "tags": [
          "Chat"
        ],
        "x-gram": {
          "name": "send_chat_message",
          "description": "<context>\nThis endpoint sends a message to the MCP-powered AI assistant and returns a response.\nThe AI can use various tools to help with task management, deployment analysis, and other operations.\n</context>\n<prerequisites>\n- userId parameter is required for authentication\n- Message content is required\n- Conversation history is optional but helps maintain context\n</prerequisites>\n<usage>\n- Use this to ask questions about your tasks, deployment status, or get AI assistance\n- The AI can access your task data and perform actions on your behalf\n- Session ID helps maintain conversation context across multiple messages\n</usage>\n",
          "responseFilterType": "jq"
        },
        "security": [
          {
            "UserIdParam": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              },
              "example": {
                "message": "Can you help me analyze the current deployment status?",
                "conversation_history": [
                  {
                    "role": "user",
                    "content": "Hello",
                    "timestamp": "2025-07-31T20:00:00.000Z"
                  },
                  {
                    "role": "assistant",
                    "content": "Hi! How can I help you today?",
                    "timestamp": "2025-07-31T20:00:05.000Z"
                  }
                ],
                "session_id": "445cf28e-b95c-46d5-8af4-c327fe33149b"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully processed chat message",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                },
                "example": {
                  "response": "I can help you analyze the deployment status. Let me check the current system status using the available tools.",
                  "session_id": "445cf28e-b95c-46d5-8af4-c327fe33149b",
                  "user_message_count": 2,
                  "tool_usage": {
                    "has_tools": true,
                    "tool_calls": [
                      {
                        "function": {
                          "name": "deployment_status_check",
                          "arguments": "{\"environment\": \"production\"}"
                        },
                        "content": [
                          {
                            "type": "text",
                            "text": "All services are running normally. Last deployment: 2025-07-31T18:30:00Z"
                          }
                        ]
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing message",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "message": "Message is required"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "message": "Unauthorized"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "message": "Failed to process chat request"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 2000
          },
          "conversation_history": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "role": {
                  "type": "string",
                  "enum": [
                    "user",
                    "assistant"
                  ]
                },
                "content": {
                  "type": "string"
                },
                "timestamp": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          },
          "session_id": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "description": "User ID (optional if provided via Bearer token or query param)"
          }
        }
      },
      "ChatResponse": {
        "type": "object",
        "required": [
          "response",
          "session_id",
          "user_message_count"
        ],
        "properties": {
          "response": {
            "type": "string"
          },
          "session_id": {
            "type": "string"
          },
          "user_message_count": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "tool_usage": {
            "type": "object",
            "nullable": true
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message"
          }
        }
      }
    },
    "securitySchemes": {
      "UserIdParam": {
        "type": "apiKey",
        "in": "query",
        "name": "userId",
        "description": "User ID parameter for authentication"
      }
    }
  }
}